# engine/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Policy rehearsal engine",
    "description": "Policy PDF \u2192 PolicyIR \u2192 Person 3 fuzz (\u226450 agents) + optional MiroFish swarm \u2192 EvaluationReport + PolicyEffectivenessReport (0\u2013100).",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/runs": {
      "post": {
        "summary": "Create Run",
        "operationId": "create_run_v1_runs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_run_v1_runs_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Runs",
        "operationId": "list_runs_v1_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs V1 Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/ir": {
      "get": {
        "summary": "Get Ir",
        "operationId": "get_ir_v1_runs__run_id__ir_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyIR"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/revise": {
      "post": {
        "summary": "Revise Run",
        "operationId": "revise_run_v1_runs__run_id__revise_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviseBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/scenarios": {
      "post": {
        "summary": "Attach Scenarios",
        "description": "Person 3 posts their ScenarioSuite here.",
        "operationId": "attach_scenarios_v1_runs__run_id__scenarios_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioSuite"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Scenarios",
        "operationId": "get_scenarios_v1_runs__run_id__scenarios_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioSuite"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/mirofish": {
      "get": {
        "summary": "Get Mirofish Pack",
        "operationId": "get_mirofish_pack_v1_runs__run_id__mirofish_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MiroFishPack"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Build Or Launch Mirofish",
        "description": "Rebuild the MiroFish seed/prompt. Pass launch=true to POST into a running MiroFish.",
        "operationId": "build_or_launch_mirofish_v1_runs__run_id__mirofish_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "launch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Launch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/mirofish/seed.md": {
      "get": {
        "summary": "Get Mirofish Seed",
        "operationId": "get_mirofish_seed_v1_runs__run_id__mirofish_seed_md_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/evaluation": {
      "get": {
        "summary": "Get Evaluation",
        "operationId": "get_evaluation_v1_runs__run_id__evaluation_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/effectiveness": {
      "get": {
        "summary": "Get Effectiveness",
        "operationId": "get_effectiveness_v1_runs__run_id__effectiveness_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyEffectivenessReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/rehearse": {
      "post": {
        "summary": "Rehearse Run",
        "description": "Person 3: grade fuzz agents and optionally run the MiroFish swarm, then score 0\u2013100.",
        "operationId": "rehearse_run_v1_runs__run_id__rehearse_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "swarm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Swarm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_run_v1_runs_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File",
            "description": "Policy PDF, TXT, or MD"
          },
          "policy_text": {
            "type": "string",
            "title": "Policy Text",
            "description": "Paste a policy if you are not uploading a file",
            "default": ""
          },
          "seed_text": {
            "type": "string",
            "title": "Seed Text",
            "default": ""
          },
          "population_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population Size"
          },
          "groups": {
            "type": "string",
            "title": "Groups",
            "description": "Comma-separated audience groups",
            "default": ""
          },
          "audience_json": {
            "type": "string",
            "title": "Audience Json",
            "description": "Structured audience segments JSON: [{\"id\",\"label\",\"weight\",\"attributes\"}] or {segments:[...]}",
            "default": ""
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "description": "Optional locale / jurisdiction label for the audience",
            "default": ""
          },
          "seed_file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed File",
            "description": "Optional audience seed PDF/TXT/MD. Text is merged into seed_text."
          }
        },
        "type": "object",
        "title": "Body_create_run_v1_runs_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MiroFishPack": {
        "properties": {
          "seed_filename": {
            "type": "string",
            "title": "Seed Filename",
            "default": "mirofish_seed.md"
          },
          "seed_markdown": {
            "type": "string",
            "title": "Seed Markdown"
          },
          "simulation_requirement": {
            "type": "string",
            "title": "Simulation Requirement"
          },
          "population_size": {
            "type": "integer",
            "title": "Population Size",
            "default": 0
          },
          "agent_count": {
            "type": "integer",
            "title": "Agent Count",
            "default": 0
          },
          "launch": {
            "$ref": "#/components/schemas/MiroFishLaunch"
          },
          "built_at": {
            "type": "string",
            "title": "Built At"
          }
        },
        "type": "object",
        "required": [
          "seed_markdown",
          "simulation_requirement"
        ],
        "title": "MiroFishPack",
        "description": "Handoff from this engine to MiroFish: a population seed + sim prompt."
      },
...
      "PolicyIR": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "source": {
            "$ref": "#/components/schemas/PolicyDocument"
          },
          "actor_types": {
            "items": {
              "$ref": "#/components/schemas/ActorType"
            },
            "type": "array",
            "title": "Actor Types"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSpec"
            },
            "type": "array",
            "title": "Actions"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "type": "array",
            "title": "Rules"
          },
          "glossary": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Glossary"
          },
          "open_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open Questions"
          },
          "conflicts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "index": {
            "$ref": "#/components/schemas/CompiledIndex"
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 1
          },
          "parent_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Revision"
          },
          "change_log": {
            "items": {
              "$ref": "#/components/schemas/ChangeLogEntry"
            },
            "type": "array",
            "title": "Change Log"
          },
          "compiled_at": {
            "type": "string",
            "title": "Compiled At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "source"
        ],
        "title": "PolicyIR",
        "description": "Compiled policy. This is the handoff from person 2 to person 3."
      },
...
      "ReviseBody": {
        "properties": {
          "instruction": {
            "type": "string",
            "title": "Instruction"
          }
        },
        "type": "object",
        "required": [
          "instruction"
        ],
        "title": "ReviseBody"
      },
...
      "RunRecord": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "ingesting",
              "extracting",
              "compiling",
              "compiled",
              "generating_scenarios",
              "scenarios_ready",
              "evaluating",
              "rehearsing",
              "completed",
              "failed"
            ],
            "title": "Status",
            "default": "pending"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "seed": {
            "$ref": "#/components/schemas/SeedSpec"
          },
          "document": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyDocument"
              },
              {
                "type": "null"
              }
            ]
          },
          "ir": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyIR"
              },
              {
                "type": "null"
              }
            ]
          },
          "suite": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioSuite"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "effectiveness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyEffectivenessReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "mirofish": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MiroFishPack"
              },
              {
                "type": "null"
              }
            ]
          },
          "extra": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extra"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "RunRecord"
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "policy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          },
          "policy_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Revision"
          },
          "rule_count": {
            "type": "integer",
            "title": "Rule Count",
            "default": 0
          },
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count",
            "default": 0
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status",
          "message"
        ],
        "title": "RunSummary"
      },
...
      "ScenarioSuite": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "1.0.0"
          },
          "suite_id": {
            "type": "string",
            "title": "Suite Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "policy_revision": {
            "type": "integer",
            "title": "Policy Revision"
          },
          "seed": {
            "type": "string",
            "title": "Seed"
          },
          "population_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Population Size"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/Scenario"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "policy_revision",
          "seed"
        ],
        "title": "ScenarioSuite",
        "description": "Person 3 output: PolicyIR + seed \u2192 ScenarioSuite."
      },
...
    }
  }
}